+++ OverseasShopping/app/src/main/java/com/example/overseasshopping/ProductFragment.java
@@ -106,7 +106,7 @@ public class ProductFragment extends Fragment {
 
             @Override
             public void onTextChanged(CharSequence s, int start, int before, int count) {
-                mProduct.setPrice(Integer.parseInt(s.toString()));
+                mProduct.setPrice(Double.parseDouble(s.toString()));
             }
 
             @Override
@@ -160,7 +160,7 @@ public class ProductFragment extends Fragment {
                 //User user = new User("ki", "ki", "la", "la", 23, 23, null);
 
                 //db.addUser(user);
-                Product product = new Product(mProductName.getText().toString(), "Empty", 1, mProductDescription.getText().toString(), Integer.parseInt(mProductPrice.getText().toString()), Integer.parseInt(mProductQuantity.getText().toString()));
+                Product product = new Product(mProductName.getText().toString(), "Empty", 1, mProductDescription.getText().toString(), Double.parseDouble(mProductPrice.getText().toString()), Integer.parseInt(mProductQuantity.getText().toString()));
                 db.addProduct(product, product.getUserNo());
                 FragmentManager fm = getFragmentManager();
                 fm.beginTransaction().replace(R.id.main_container, new ProductListFragment()).commit();
